% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_search.R
\name{item_search}
\alias{item_search}
\title{Search for Product on Walmart.com}
\usage{
item_search(apikey, query, categoryID = NULL, start = NULL, sort = NULL,
  order = NULL, numItems = 10, facet = NULL, facet_filter = NULL,
  facet_range = NULL)
}
\arguments{
\item{apikey}{Your API access key authorized by Walmart API (Required)}

\item{query}{Search text - whitespace separated sequence of keywords to search for (Required)}

\item{categoryID}{Category id of the category for search within a category. This should match the id field from Taxonomy API (see function taxonomy{walmartR}) (Optional)}

\item{start}{Starting point of the results within the matching set of items - upto 10 items will be returned starting from this item (Optional)}

\item{sort}{Sorting criteria, allowed sort types are [relevance, price, title, bestseller, customerRating, new]. Default sort is by relevance (Optional)}

\item{order}{Sort ordering criteria, allowed values are [asc, desc]. This parameter is needed only for the sort types [price, title, customerRating].(Optional)}

\item{numItems}{Number of matching items to be returned, max value 25. Default is 10 (Optional)}

\item{facet}{Boolean flag to enable facets. Default value is off. Set this to on to enable facets  (Optional)}

\item{facet_filter}{Filter on the facet attribute values. This parameter can be set to <facet-name>:<facet-value> (without the angles). Here facet-name and facet-value can be any valid facet picked from the search API response when facets are on (Optional)}

\item{facet_range}{Range filter for facets which take range values, like price. See usage above in the examples (Optional)}
}
\value{
Returns an RDS file named "item_search_output.RDS".
}
\description{
This function allows text search on the Walmart.com catalogue and returns matching items available for sale online.
}
\examples{
apikey = "xs9emg846j69q7e9fhbfxz6d"
item_search(apikey, query = "tv", facet = "on", facet_filter = "brand:Samsung")

}
\references{
https://developer.walmartlabs.com/docs/read/Search_API
}
