% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postbrowse_recommendation.R
\name{postbrowse_recommendation}
\alias{postbrowse_recommendation}
\title{Post Browsed Products Recommendation}
\usage{
postbrowse_recommendation(apikey = NULL, itemId)
}
\arguments{
\item{apikey}{Your API access key authorized by Walmart API (Required)}

\item{itemId}{The item id associated with an item on Walmart.com. This can be generally seen in the last 8 digits of a URL string, but can also be acquired by using the product_lookup function (Required)}
}
\value{
Returns an RDS file named "postbrowse_recommendation_output.RDS".
}
\description{
This function allows you to recommend products to someone based on their product viewing history and returns top 10 post browsed products for an item.
}
\examples{
apikey = "YOUR API KEY"
postbrowse_recommendation(apikey, "36904791")

}
\references{
https://developer.walmartlabs.com/docs/read/Post_Browsed_Products_API
}
