% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_lookup.R
\name{product_lookup}
\alias{product_lookup}
\title{Lookup product information on Walmart}
\usage{
product_lookup(apikey, ids)
}
\arguments{
\item{apikey}{Your API access key authorized by Walmart API (Required)}

\item{ids}{Comma separated list of item ids (Re)}
}
\value{
Returns an RDS file named "product_lookup_output.RDS".
}
\description{
This function gives access to item price and availability in real-time. Item id needs to be specified for doing a lookup.
}
\examples{
apikey = "xs9emg846j69q7e9fhbfxz6d"
product_lookup(apikey, "12417832")
product_lookup(apikey, "12417832,19336123")

}
\references{
https://developer.walmartlabs.com/docs/read/Home
}
