% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_locator.R
\name{store_locator}
\alias{store_locator}
\title{Store Locator}
\usage{
store_locator(apikey, city = NULL, lat = NULL, lon = NULL, zip = NULL)
}
\arguments{
\item{apikey}{Your API access key authorized by Walmart API (Required)}

\item{city}{City name, capitalized (Optional)}

\item{lat}{Latitude (Optional)}

\item{lon}{Longitude (Optional)}

\item{zip}{Zip (Optional)}
}
\value{
Returns an RDS file named "store_locations_output.RDS".
}
\description{
This function helps locate nearest Walmart Stores via API. It lets users search for stores by latitude and longitude, by zip code and by city.
}
\examples{
apikey = "xs9emg846j69q7e9fhbfxz6d"
store_locator(apikey, city = "NEW YORK")

}
\references{
https://developer.walmartlabs.com/docs/read/Store_Locator_API
}
