% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Walmart.com Category Taxonomy}
\usage{
taxonomy(apikey)
}
\arguments{
\item{apikey}{Your API access key authorized by Walmart API (Required)}
}
\value{
Returns an RDS file named "taxonomy_output.RDS".
}
\description{
This function exposes the category taxonomy used by walmart.com to categorize items.
}
\examples{
apikey = "xs9emg846j69q7e9fhbfxz6d"
taxonomy(apikey)

}
\references{
https://developer.walmartlabs.com/docs/read/Taxonomy_API
}
